+++ modules/@apostrophecms/color-field/ui/apos/lib/AposColorSaturation.vue
@@ -73,7 +73,6 @@ export default {
       this.$emit('change', param);
     },
     handleMouseDown(e) {
-      // this.handleChange(e, true)
       window.addEventListener('mousemove', this.handleChange);
       window.addEventListener('mouseup', this.handleChange);
       window.addEventListener('mouseup', this.handleMouseUp);