+++ modules/@apostrophecms/color-field/ui/apos/logic/AposInputColor.js
@@ -76,7 +76,11 @@ export default {
     },
     update(value) {
       this.tinyColorObj = new TinyColor(value.hsl);
-      this.next = this.tinyColorObj.toString(this.format);
+      if (value._cssVariable) {
+        this.next = value._cssVariable;
+      } else {
+        this.next = this.tinyColorObj.toString(this.format);
+      }
     },
     validate(value) {
       if (this.field.required) {
@@ -90,7 +94,9 @@ export default {
       }
 
       const color = new TinyColor(value);
-      return color.isValid ? false : 'Error';
+      if (!value.startsWith('--')) {
+        return color.isValid ? false : 'Error';
+      }
     },
     clear() {
       this.next = '';