+++ modules/@apostrophecms/schema/lib/addFieldTypes.js
@@ -751,7 +751,8 @@ module.exports = (self) => {
       {
         fetchRelationships = true,
         ancestors = [],
-        isParentVisible = true
+        rootConvert = true,
+        ancestorPath = ''
       } = {}
     ) {
       const schema = field.schema;
@@ -777,7 +778,8 @@ module.exports = (self) => {
           const options = {
             fetchRelationships,
             ancestors: [ ...ancestors, destination ],
-            isParentVisible
+            rootConvert,
+            ancestorPath
           };
           await self.convert(req, schema, datum, result, options);
         } catch (e) {
@@ -867,7 +869,8 @@ module.exports = (self) => {
       {
         fetchRelationships = true,
         ancestors = {},
-        isParentVisible = true,
+        rootConvert = true,
+        ancestorPath = '',
         doc = {}
       } = {}
     ) {
@@ -881,7 +884,8 @@ module.exports = (self) => {
       const options = {
         fetchRelationships,
         ancestors: [ ...ancestors, destination ],
-        isParentVisible
+        rootConvert,
+        ancestorPath
       };
       if (data == null || typeof data !== 'object' || Array.isArray(data)) {
         data = {};
@@ -978,12 +982,14 @@ module.exports = (self) => {
       destination,
       {
         fetchRelationships = true,
-        isParentVisible = true
+        rootConvert = true,
+        ancestorPath = ''
       } = {}
     ) {
       const options = {
         fetchRelationships,
-        isParentVisible
+        rootConvert,
+        ancestorPath
       };
       const manager = self.apos.doc.getManager(field.withType);
       if (!manager) {
@@ -1067,15 +1073,16 @@ module.exports = (self) => {
           if (result) {
             actualDocs.push(result);
           }
-        } else if ((item && ((typeof item._id) === 'string'))) {
+        } else if ((item && (typeof item._id === 'string'))) {
           const result = results.find(doc => (doc._id === item._id));
           if (result) {
             if (field.schema) {
+              const destItem = (Array.isArray(destination[field.name]) ? destination[field.name] : [])
+                .find((doc) => doc._id === item._id);
               result._fields = {
-                ...(destination[field.name]
-                  ?.find?.(doc => doc._id === item._id)
-                  ?._fields || {})
+                ...destItem?._fields || {}
               };
+
               if (item && ((typeof item._fields === 'object'))) {
                 await self.convert(req, field.schema, item._fields || {}, result._fields, options);
               }