+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Focus properly Widget Editor modals when opened. Keep the previous active focus on the modal when closing the widget editor.
 * a11y improvements for context menus.
+* Fixes broken widget preview URL when the image is overridden (module improve) and external build module is registered.
 
 ## 4.10.0 (2024-11-20)