+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Adds support for supplying CSS variable names to a color field's `presetColors` array as selectable values.
 * Adds support for dynamic focus trap in Context menus (prop `dynamicFocus`). When set to `true`, the focusable elements are recalculated on each cycle step.
 * Adds option to disable `tabindex` on `AposToggle` component. A new prop `disableFocus` can be set to `false` to disable the focus on the toggle button. It's enabled by default.
+* Adds new prop `additionalMenuItems` to `AposDocContextMenu`. Allow to pass additional menu items that can trigger an action when clicked, or simply pass the action as the close event parameter.
 
 ## 4.10.0 (2024-11-20)