+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Fixes broken widget preview URL when the image is overridden (module improve) and external build module is registered.
 * Inject dynamic custom bundle CSS when using external build module with no CSS entry point.
 * Range field now correctly takes 0 into account.
+* Apos style does not go through `postcss-viewport-to-container-toggle` plugin anymore to avoid UI bugs.
 
 ### Adds