+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 * Focus properly Widget Editor modals when opened. Keep the previous active focus on the modal when closing the widget editor.
 * a11y improvements for context menus.
 * Fixes broken widget preview URL when the image is overridden (module improve) and external build module is registered.
-* In th schema convert method, we wait all sub convert to run, to have access to the final destination object in order to check if 
+* In the schema convert method, we wait all sub convert to run, to have access to the final destination object in order to check if 
 fields (or their parents) are visible and so relates errors discarded.
 
 ### Adds